Extract ModelOnDisk to its own module
