+++ scripts/strip_models.py
@@ -22,7 +22,7 @@ from pathlib import Path
 import humanize
 import torch
 
-from invokeai.backend.model_manager.config import FSLayout, ModelOnDisk
+from invokeai.backend.model_manager.model_on_disk import ModelOnDisk
 from invokeai.backend.model_manager.search import ModelSearch
 
 
@@ -62,7 +62,7 @@ def load_stripped_model(path: Path, *args, **kwargs):
 
 def create_stripped_model(original_model_path: Path, stripped_model_path: Path) -> ModelOnDisk:
     original = ModelOnDisk(original_model_path)
-    if original.layout == FSLayout.FILE:
+    if original.path.is_file():
         shutil.copy2(original.path, stripped_model_path)
     else:
         shutil.copytree(original.path, stripped_model_path, dirs_exist_ok=True)
